# Optimize queries using regex matchers for set lookups