--- a/querier_test.go
+++ b/querier_test.go
@@ -21,7 +21,6 @@ import (
 	"os"
 	"path/filepath"
 	"sort"
-	"strconv"
 	"testing"
 
 	"github.com/pkg/errors"