fixes https://github.com/prometheus/prometheus/issues/2651