@brian-brazil 2019-05-14
You need to check for and exclude the wrapper we put around the regex the user specifies.