@brian-brazil 2019-05-14
If the wrapper is missing, treat it as a regex.