@brian-brazil 2019-05-14
At some point, us doing parsing of the strings might be slower than the regex. Would be good to check one way or the other.