@bboreham 2019-05-14
What's the motivation for only allowing escaping of special characters?  What if I want to match `\t` or `\n` in a label value?