@codesome 2019-05-14
Adding to @bboreham's comment, isn't it enough to just include any character that is escaped rather than checking for special character or '\\'?