@brian-brazil 2019-05-15
What about 100k to 1M series? It's the bigger data sizes where the wins are, what you've tested are already fast.