@gouthamve 2019-05-15
This looks good so far!

Can you also add a `nBlocks=10` version for `BenchmarkSetMatcher/nSeries=500000,nBlocks=1,pattern="^(?:1|2|3|4|5|6|7|8|9|10)$"-8`?