@naivewong 2019-05-15
Unable to show you the results because the previous one was already the limit of my MBP.