@krasi-georgiev 2019-05-15
pinged you on irc to give you access to a big machine where you can run these test quickly.